fix(scripts): compile wizard installer from extracted app bundle tar
electrobun deletes the app bundle folder after tarring (buildEnvironment
!== "dev"), leaving only the self-extracting wrapper (extractor stub +
inner tar.zst) at build/<channel>-win-x64/LLMSpace-<channel>/; the wizard
hook compiled ISCC from the wrapper, so the installer shipped a bundle
without bin/launcher.exe and failed to launch. At postPackage the only
complete copy is the channel tar.zst: decompress it (helpers moved to a
shared win-tar module), strip the tar's bundle-folder root component so
{app}\bin\launcher.exe lands flat, compile ISCC from a dot-prefixed
staging dir, and clean up after. Channels without a tar.zst (dev) skip.
